requirements.txt
pillow>=10.0
pypdf>=4.0
PyMuPDF>=1.23
PyYAML>=6.0

